refactor: reorganize repository
