+++ customizers/addBabelPlugins.js
@@ -1,0 +1,5 @@
+const addBabelPlugin = require("./addBabelPlugin");
+
+const addBabelPlugins = (...plugins) => plugins.map(p => addBabelPlugin(p));
+
+module.exports = addBabelPlugins;
