+++ customizers/addBabelPreset.js
@@ -1,0 +1,8 @@
+const getBabelLoader = require("../utilities/getBabelLoader");
+
+const addBabelPreset = preset => config => {
+  getBabelLoader(config).options.presets.push(preset);
+  return config;
+};
+
+module.exports = addBabelPreset;
