+++ customizers/addBabelPresets.js
@@ -1,0 +1,5 @@
+const addBabelPreset = require("./addBabelPreset");
+
+const addBabelPresets = (...plugins) => plugins.map(p => addBabelPreset(p));
+
+module.exports = addBabelPresets;
