+++ customizers/addBundleVisualizer.js
@@ -1,0 +1,23 @@
+const addBundleVisualizer = (options = {}, behindFlag = false) => config => {
+  const BundleAnalyzerPlugin = require("webpack-bundle-analyzer")
+    .BundleAnalyzerPlugin;
+
+  // if behindFlag is set to true, the report will be created only if
+  // the `--analyze` flag is added to the `yarn build` command
+  if (behindFlag ? process.argv.includes("--analyze") : true) {
+    config.plugins.push(
+      new BundleAnalyzerPlugin(
+        Object.assign(
+          {
+            analyzerMode: "static",
+            reportFilename: "report.html"
+          },
+          options
+        )
+      )
+    );
+  }
+  return config;
+};
+
+module.exports = addBundleVisualizer;
