+++ customizers/addWebpackAlias.js
@@ -1,0 +1,12 @@
+const addWebpackAlias = alias => config => {
+  if (!config.resolve) {
+    config.resolve = {};
+  }
+  if (!config.resolve.alias) {
+    config.resolve.alias = {};
+  }
+  Object.assign(config.resolve.alias, alias);
+  return config;
+};
+
+module.exports = addWebpackAlias;
