+++ customizers/adjustWorkbox.js
@@ -1,0 +1,10 @@
+const adjustWorkbox = adjust => config => {
+  config.plugins.forEach(p => {
+    if (p.constructor.name === "GenerateSW") {
+      adjust(p.config);
+    }
+  });
+  return config;
+};
+
+module.exports = adjustWorkbox;
