+++ customizers/disableEslint.js
@@ -1,0 +1,11 @@
+const disableEsLint = () => config => {
+  let eslintRules = config.module.rules.filter(
+    r => r.use && r.use.some(u => u.options && u.options.useEslintrc !== void 0)
+  );
+  eslintRules.forEach(rule => {
+    config.module.rules = config.module.rules.filter(r => r !== rule);
+  });
+  return config;
+};
+
+module.exports = disableEsLint;
