+++ customizers/fixBabelImports.js
@@ -1,0 +1,16 @@
+const addBabelPlugin = require("./addBabelPlugin");
+
+const fixBabelImports = (libraryName, options) =>
+  addBabelPlugin([
+    "import",
+    Object.assign(
+      {},
+      {
+        libraryName
+      },
+      options
+    ),
+    `fix-${libraryName}-imports`
+  ]);
+
+module.exports = fixBabelImports;
