+++ customizers/useBabelRc.js
@@ -1,0 +1,8 @@
+const getBabelLoader = require("../utilities/getBabelLoader");
+
+const useBabelRc = () => config => {
+  getBabelLoader(config).options.babelrc = true;
+  return config;
+};
+
+module.exports = useBabelRc;
