+++ customizers/useEslintRc.js
@@ -1,0 +1,22 @@
+const useEslintRc = configFile => config => {
+  const eslintRule = config.module.rules.filter(
+    r => r.use && r.use.some(u => u.options && u.options.useEslintrc !== void 0)
+  )[0];
+
+  eslintRule.use[0].options.useEslintrc = true;
+  eslintRule.use[0].options.ignore = true;
+  eslintRule.use[0].options.configFile = configFile;
+
+  delete eslintRule.use[0].options.baseConfig;
+
+  const rules = config.module.rules.map(r =>
+    r.use && r.use.some(u => u.options && u.options.useEslintrc !== void 0)
+      ? eslintRule
+      : r
+  );
+  config.module.rules = rules;
+
+  return config;
+};
+
+module.exports = useEslintRc;
