+++ package.json
@@ -2,14 +2,29 @@
   "name": "customize-cra",
   "version": "0.2.12",
   "description": "",
+  "author": "Adam Rackis",
+  "license": "MIT",
   "repository": "arackaf/customize-cra",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "prettier": "pretty-quick --staged",
+    "eslint": "eslint ."
   },
-  "author": "Adam Rackis",
-  "license": "MIT",
+  "husky": {
+    "hooks": {
+      "pre-commit": "npm run prettier && npm run eslint --fix"
+    }
+  },
   "dependencies": {
-    "lodash.flow": "^3.5.0"
+    "lodash.flow": "^3.5.0",
+    "require-all": "^3.0.0"
+  },
+  "devDependencies": {
+    "babel-eslint": "^10.0.1",
+    "eslint": "^5.15.3",
+    "husky": "^1.3.1",
+    "prettier": "^1.16.4",
+    "pretty-quick": "^1.10.0"
   }
 }
